+++ src/main/java/ladysnake/effective/client/RenderedHypnoEntities.java
@@ -1,0 +1,14 @@
+package ladysnake.effective.client;
+
+import net.minecraft.entity.passive.AxolotlEntity;
+import net.minecraft.entity.passive.GlowSquidEntity;
+
+import java.util.HashSet;
+import java.util.Set;
+
+public class RenderedHypnoEntities {
+    public static Set<GlowSquidEntity> GLOWSQUIDS = new HashSet<>();
+    public static double lookIntensity = 0f;
+    public static double lookIntensityGoal = 0f;
+    public static int lockedIntensityTimer = 0;
+}