+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Effective - Changelog:
 
+## Effective 1.4 - 1.19.2
+- Hypnotizing glow squids
+  - Glow squids can now hypnotize you, displaying a hypnotizing shader that gradually gets stronger the longer you look at them
+  - Glow squids will attract your cursor if they appear on your screen, this attraction getting stronger with the shader
+  - Both cursor attraction and hypnotizing can be disabled with two new configuration options
+
 ## Effective 1.3 - 1.19.2
 - Overhauled splashes:
   - Splashes will now be colored depending on the water they originate from